+++ src/pkg/datastore/Makefile
@@ -1,6 +1,6 @@
 include $(GOROOT)/src/Make.inc
 
-TARG=Shock/DataStore
+TARG=shock/datastore
 GOFILES=\
 	Node.go Store.go
 include $(GOROOT)/src/Make.pkg
